refactor: UIViewController을 extension하여 삭제 시 보여주는 얼럿의 재사용성을 개선
