+++ Diary/Resource/Controller/Extension/UIViewController+.swift
@@ -1,0 +1,25 @@
+//
+//  UIViewController+.swift
+//  Diary
+//
+//  Created by SummerCat and som on 2022/12/29.
+//
+
+import UIKit
+
+extension UIViewController {
+    func showDeleteAlert(for diaryModel: DiaryModel?, handler: @escaping ((UIAlertAction) -> Void)) {
+        let alert: UIAlertController = UIAlertController(title: Namespace.deleteDiary,
+                                                         message: Namespace.deleteMessage,
+                                                         preferredStyle: .alert)
+        let cancelAction: UIAlertAction = UIAlertAction(title: Namespace.cancel,
+                                                        style: .cancel)
+        let deleteAction: UIAlertAction = UIAlertAction(title: Namespace.delete,
+                                                        style: .destructive,
+                                                        handler: handler)
+        
+        alert.addAction(cancelAction)
+        alert.addAction(deleteAction)
+        present(alert, animated: true)
+    }
+}
